Refactored SQLCreateInterpreter.
In order to remove the two lists I changed the code that visits constraints
to push to the stack a function responsible for applying the constraint at
a later time (when the column is already added to the table).
The only exceptions are the column's IDENTITY and DEFAULT. This constraints
are applied when visiting the ColumnDefinition.