Simplified previous commit by setting primary keys directly.
It seems DataTable.PrimaryKey takes care of configuring the constraints so
I don't have to worry about applying the foreign keys last. However, I do
need to check that the primary key hasn't been set before in order to
throw an error. Also, if the unique constraint is not a PK and it has been
set before I simply update its name instead of throwing an error.