Changed what visiting a selectExpression returns
Since i can have kind of anything there, i changed it from returning the column to returning a name, and a function that given a row returns the value that column should adopt for that row.

This is used to construct the recordset.

This breaks down all the other operations, see the next commit for that