Implemented comparison of integer compatible types.
I added SQLComparer. It's kind of a hack. What it does is it tries to
convert both values to integers. If the conversion is successful it
compares the integers. Otherwise it delegates to the default comparer.
We could probably implement this better, but for now it allows to
compare integer values of different types (including strings that can be
parsed to integer).